修正文件超大时整形数据类型计算异常 Signed-off-by: xxonehjh <code234@qq.com>

+++ src/main/java/com/hjh/files/sync/common/RemoteFileManage.java
@@ -6,6 +6,6 @@ public interface RemoteFileManage {
 
 	public String md5(String file);
 
-	public byte[] part(String file, int part ,int part_size);
+	public byte[] part(String file, long part ,long part_size);
 
 }
